refactor(mxfp4-moe): dissolve kernels/utils.py into moegemm (kernel-specific)
kernels/utils.py was 100% MoE-specific (created by the MoE consolidation, only
imported by the MoE files). A generic name for MoE config + kernel geometry +
quant helpers is misleading, so fold it all into moegemm.py where it's used:
- shape constants (NE/TOPK/H/INTER defaults, BN/BK/KH_TILE/kStages/...),
  K-derived size formulas, raw/pointer/LDS helpers, e8m0/SwiGLU math.
- inline the two identical *_c_k1_for wrappers (0 external users).
- moe_dispatcher now imports these from .moegemm; decouple fp8_gemm_utils by
  taking `raw` from flydsl.expr.arith (_to_raw) instead of the MoE module.
- delete kernels/utils.py.
Pure move (no logic change): cos unchanged (fp4 0.9882 / a8w4 0.9996); ruff/black clean.
